\name{euclidean-package}
\alias{euclidean-package}
\alias{euclidean}
\docType{package}
\title{
\packageTitle{euclidean}
}
\description{
\packageDescription{euclidean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{euclidean}
\packageIndices{euclidean}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{euclidean}

Maintainer: \packageMaintainer{euclidean}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
