\name{Euclidean-package}
\alias{Euclidean-package}
\alias{Euclidean}
\docType{package}
\title{
\packageTitle{Euclidean}
}
\description{
\packageDescription{Euclidean}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Euclidean}
\packageIndices{Euclidean}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{Euclidean}

Maintainer: \packageMaintainer{Euclidean}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
